git-redo: Add (non-interactive amend tool)
`git-redo` is inspired by `git commit --amend`, with the goal of allowing
to edit any commit on the current branch - and not only the latest one.

One of the most important characteristics of `git-redo` is that it runs
non-interactively - i.e. the commit to be edited needs to be provided on
the command line - the traditional todo sequence editor is not shown.